Separation of logic. All methods for creating objects placed in CommandProcessor class.
